Bird point count data from Alberta's Oil Sands region

Data set used in Mahon et al. 2016 and Mahon et al. 2019.

Usage

data("josm")

Format

A list with elements surveys, species, and counts.

Source

Mahon et al. 2016 and Mahon et al. 2019.

References

Mahon, C. L., Holloway, G., Solymos, P., Cumming, S. G., Bayne, E. M., Schmiegelow, F. K. A., Song, S. J., 2016. Community structure and niche characteristics of upland and lowland western boreal birds at multiple spatial scales Forest Ecology and Management 361:99-116. <doi:10.1016/j.foreco.2015.11.007>

Mahon, C. L., Holloway, G., Bayne, E. M., Toms, J. D., 2019. Additive and interactive cumulative effects on boreal landbirds: winners and losers in a multi-stressor landscape Ecological Applications 29:e01895. <doi:10.1002/eap.1895>

Examples

data(josm)
str(josm)
#> List of 3
#>  $ surveys:'data.frame':	4569 obs. of  36 variables:
#>   ..$ SiteID       : Factor w/ 4569 levels "CL10102","CL10106",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ SurveyArea   : Factor w/ 271 levels "CL101","CL102",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Longitude    : num [1:4569] -113 -113 -113 -113 -113 ...
#>   ..$ Latitude     : num [1:4569] 55.2 55.2 55.2 55.2 55.2 ...
#>   ..$ Date         : Date[1:4569], format: "2013-06-22" "2013-06-22" ...
#>   ..$ StationID    : Factor w/ 4569 levels "CL10102-1","CL10106-1",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ ObserverID   : int [1:4569] 82 48 48 82 48 48 48 82 82 82 ...
#>   ..$ TimeStart    : Factor w/ 377 levels "10:00:00 AM",..: 76 153 75 116 239 200 333 277 161 309 ...
#>   ..$ VisitID      : int [1:4569] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ WindStart    : int [1:4569] 0 0 0 0 0 1 2 0 0 0 ...
#>   ..$ PrecipStart  : int [1:4569] 0 0 0 0 0 2 0 0 0 0 ...
#>   ..$ TempStart    : num [1:4569] 8 4 4 8 9 8 13 11 8 13 ...
#>   ..$ CloudStart   : int [1:4569] 0 3 0 0 3 3 0 2 3 1 ...
#>   ..$ WindEnd      : int [1:4569] 0 0 0 0 0 1 2 0 0 2 ...
#>   ..$ PrecipEnd    : int [1:4569] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ TempEnd      : num [1:4569] 7 4 4 7 9 8 14 10 7 13 ...
#>   ..$ CloudEnd     : int [1:4569] 0 3 0 0 3 3 0 2 3 1 ...
#>   ..$ TimeFin      : Factor w/ 370 levels "10:00:00 AM",..: 87 164 86 127 250 211 344 288 172 320 ...
#>   ..$ Noise        : int [1:4569] 1 0 0 1 0 0 1 0 0 1 ...
#>   ..$ OvernightRain: Factor w/ 2 levels "FALSE","TRUE": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ DateTime     : POSIXct[1:4569], format: "2013-06-22 05:12:00" "2013-06-22 06:29:00" ...
#>   ..$ SunRiseTime  : POSIXct[1:4569], format: "2013-06-22 04:52:59" "2013-06-22 04:53:04" ...
#>   ..$ SunRiseFrac  : num [1:4569] 0.203 0.204 0.204 0.203 0.204 ...
#>   ..$ TSSR         : num [1:4569] 0.0132 0.0666 0.0125 0.041 0.1263 ...
#>   ..$ OrdinalDay   : int [1:4569] 172 172 172 172 172 172 172 172 172 172 ...
#>   ..$ DAY          : num [1:4569] 0.471 0.471 0.471 0.471 0.471 ...
#>   ..$ Open         : num [1:4569] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ Water        : num [1:4569] 0.02055 0.00752 0.00752 0.03284 0.00416 ...
#>   ..$ Agr          : num [1:4569] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ UrbInd       : num [1:4569] 0 0 0 0.01056 0.00839 ...
#>   ..$ SoftLin      : num [1:4569] 0.00521 0.01811 0.01811 0.00141 0.01349 ...
#>   ..$ Roads        : num [1:4569] 0.052461 0.000913 0.000913 0.003307 0.046578 ...
#>   ..$ Decid        : num [1:4569] 0.8569 0.0443 0.0443 0.8823 0.8139 ...
#>   ..$ OpenWet      : num [1:4569] 0.00315 0.58355 0.58355 0.06961 0.11164 ...
#>   ..$ Conif        : num [1:4569] 0.0249 0 0 0 0 ...
#>   ..$ ConifWet     : num [1:4569] 0.0368 0.3456 0.3456 0 0.0018 ...
#>  $ species:'data.frame':	149 obs. of  3 variables:
#>   ..$ SpeciesID     : Factor w/ 150 levels "ALFL","AMBI",..: 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ SpeciesName   : Factor w/ 149 levels "Alder Flycatcher",..: 1 2 3 4 5 6 7 8 9 11 ...
#>   ..$ ScientificName: Factor w/ 145 levels "","Accipiter cooperii",..: 61 23 70 47 31 69 1 117 135 10 ...
#>  $ counts :'data.frame':	52372 obs. of  18 variables:
#>   ..$ ObservationID: Factor w/ 57024 levels "CL10102-130622-001",..: 1 2 3 4 5 6 8 9 10 11 ...
#>   ..$ SiteID       : Factor w/ 4569 levels "CL10102","CL10106",..: 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ StationID    : Factor w/ 4569 levels "CL10102-1","CL10106-1",..: 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ TimeInterval : int [1:52372] 1 1 1 1 5 5 1 1 1 1 ...
#>   ..$ Direction    : int [1:52372] 1 2 2 2 1 4 4 4 1 1 ...
#>   ..$ Distance     : int [1:52372] 1 2 2 1 3 3 2 1 1 1 ...
#>   ..$ DetectType1  : Factor w/ 3 levels "C","S","V": 2 2 2 2 1 1 2 2 2 2 ...
#>   ..$ DetectType2  : Factor w/ 3 levels "C","S","V": NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ DetectType3  : Factor w/ 3 levels "C","S","V": NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ Sex          : Factor w/ 4 levels "F","M","P","U": 2 2 2 2 4 4 2 2 2 2 ...
#>   ..$ Age          : Factor w/ 6 levels "A","F","J","JUV",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Activity1    : Factor w/ 17 levels "BE","CF","CH",..: 5 5 5 5 NA NA NA 5 5 NA ...
#>   ..$ Activity2    : Factor w/ 17 levels "48","BE","CF",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ Activity3    : Factor w/ 7 levels "CF","DC","DR",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ ActivityNote : Factor w/ 959 levels "AGITATED","AGITATED CALLING",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ Dur          : Factor w/ 3 levels "0-3min","3-5min",..: 1 1 1 1 3 3 1 1 1 1 ...
#>   ..$ Dis          : Factor w/ 3 levels "0-50m","50-100m",..: 1 2 2 1 3 3 2 1 1 1 ...
#>   ..$ SpeciesID    : Factor w/ 149 levels "ALFL","AMBI",..: 107 95 95 107 46 43 140 95 125 38 ...